More srp components added
