+++ src/components/Cards/index.js
@@ -1,0 +1,3 @@
+import Cards from './Cards.jsx';
+
+export default Cards;
