+++ src/components/Filters/index.js
@@ -1,0 +1,3 @@
+import Filters from './Filters.jsx';
+
+export default Filters;
